Start to split out CLI from recording code
