+++ spicerecord/__main__.py
@@ -1,0 +1,2 @@
+from .cli import main
+main()
